+++ src/imio/zamqp/dms/consumer.py
@@ -31,6 +31,8 @@ import tarfile
 
 cg_separator = ' ___ '
 
+# TODO add package tests
+
 # INCOMING MAILS #
 
 
@@ -255,11 +257,12 @@ class OutgoingGeneratedMail(DMSMainFile, CommonMethods):
                 self.document.outgoing_date = (self.scan_fields['scan_date'] and self.scan_fields['scan_date'] or
                                                datetime.datetime.now())
                 self.document.reindexObject(idxs=('in_out_date', ))
-            if not params['PC']:
+            if not params['PC'] and (not self.document.is_email() or self.document.email_status):
                 # close
                 trans = {
-                    'created': ['propose_to_be_signed', 'mark_as_sent'], 'scanned': [],
-                    'proposed_to_n_plus_1': ['propose_to_be_signed', 'mark_as_sent'],
+                    'created': ['mark_as_sent', 'propose_to_be_signed', 'set_to_print', 'propose_to_n_plus_1'],
+                    'scanned': ['mark_as_sent'],
+                    'proposed_to_n_plus_1': ['mark_as_sent', 'propose_to_be_signed', 'set_to_print'],
                     'to_be_signed': ['mark_as_sent'], 'to_print': ['propose_to_be_signed', 'mark_as_sent']
                 }
                 state = api.content.get_state(self.document)